pyproject.toml
[project]
name = "projeto-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.2.0",
    "beautifulsoup4>=4.14.3",
    "ipykernel>=7.2.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
]
